better installation

+++ install.sh
@@ -1,14 +1,16 @@
 echo "Creating virtual environment..."
 python3 -m venv .venv
 
-echo "Installing build-essential libdbus-glib-1-dev libgirepository1.0-dev..."
-sudo apt install build-essential libdbus-glib-1-dev libgirepository1.0-dev
+if ! dpkg -s build-essential libdbus-glib-1-dev libgirepository1.0-dev; then
+    echo "Installing build-essential libdbus-glib-1-dev libgirepository1.0-dev..."
+    sudo apt install build-essential libdbus-glib-1-dev libgirepository1.0-dev
 
 echo "Installing packages..."
 .venv/bin/python3 -m pip install -r requirements.txt
 
-echo "Running setup script..."
-.venv/bin/python3 setup.py
+if ! .venv/bin/python3 setup.py; then
+    echo "And error has occurred."
+    exit 1
 
 echo "Installing service..."
 mv .unit_service ~/.config/systemd/user/lyric-overlay.service
